Food-Review-System-main/Food-Review-System-main/requirements.txt
customtkinter
pillow
nltk